add_swagger: approach #1 fix:1
