+++ onebusaway-transit-data-federation-webapp/src/main/java/org/onebusaway/transit_data_federation_webapp/configs/SpringFoxConfig.java
@@ -3,27 +3,44 @@ package org.onebusaway.transit_data_federation_webapp.configs;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import springfox.documentation.builders.ApiInfoBuilder;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
+import springfox.documentation.service.ApiInfo;
 import springfox.documentation.spi.DocumentationType;
 import springfox.documentation.spring.web.plugins.Docket;
 import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
 @Configuration
 @EnableSwagger2
-public class SpringFoxConfig {
+@EnableWebMvc
+public class SpringFoxConfig implements WebMvcConfigurer {
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry.addResourceHandler("swagger-ui.html")
+                .addResourceLocations("classpath:/META-INF/resources/");
 
+        registry.addResourceHandler("/webjars/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/");
+    }
     @Bean
     public Docket api() {
         return new Docket(DocumentationType.SWAGGER_2)
                 .select()
-                .apis(RequestHandlerSelectors.basePackage("org.onebusaway.transit_data_federation_webapp.controllers.gtfs_realtime"))
-                .apis(RequestHandlerSelectors.basePackage("org.onebusaway.federations.webapp"))
-                .apis(RequestHandlerSelectors.basePackage("org.onebusaway.gtfs_realtime.archiver.controller"))
-                .apis(RequestHandlerSelectors.basePackage("org.onebusaway.transit_data_federation_webapp.controllers"))
-                .apis(RequestHandlerSelectors.basePackage("org.onebusaway.aws.monitoring.api"))
-                .apis(RequestHandlerSelectors.basePackage("org.onebusaway.transit_data_federation_webapp.siri"))
+                .apis(RequestHandlerSelectors.any())
                 .paths(PathSelectors.any())
+                .build()
+                .apiInfo(apiInfo());
+
+    }
+    private ApiInfo apiInfo() {
+        return new ApiInfoBuilder()
+                .title("OBA")
+                .description("Api Services")
+                .version("2.1.2")
                 .build();
     }
 }
