+++ onebusaway-transit-data-federation-webapp/src/main/webapp/WEB-INF/remoting-servlet.xml
@@ -17,10 +17,11 @@
 
 -->
 <beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns:context="http://www.springframework.org/schema/context" xmlns:aop="http://www.springframework.org/schema/aop"
-    xmlns:tx="http://www.springframework.org/schema/tx" xmlns:sec="http://www.springframework.org/schema/security"
-    xmlns:mvc="http://www.springframework.org/schema/mvc"
-    xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
+       xmlns:context="http://www.springframework.org/schema/context"
+       xmlns:aop="http://www.springframework.org/schema/aop"
+       xmlns:tx="http://www.springframework.org/schema/tx" xmlns:sec="http://www.springframework.org/schema/security"
+       xmlns:mvc="http://www.springframework.org/schema/mvc"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd
            http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context.xsd
            http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx.xsd
            http://www.springframework.org/schema/aop http://www.springframework.org/schema/aop/spring-aop.xsd
@@ -49,7 +50,14 @@
             <mvc:mapping path="/playback**.do" />
             <bean class="org.onebusaway.transit_data_federation_webapp.controllers.PlaybackInterceptor" />
         </mvc:interceptor>
+        <mvc:interceptor>
+            <mvc:mapping path="/swagger-ui/**"/>
+            <mvc:mapping path="/v2/api-docs/**"/>
+            <bean class="org.onebusaway.api.configs.SpringFoxConfig"/>
+        </mvc:interceptor>
     </mvc:interceptors>
+    <mvc:resources mapping="/swagger-ui/**" location="classpath:/META-INF/resources/webjars/springfox-swagger-ui/"/>
+
 
     <bean id="viewResolver" class="org.springframework.web.servlet.view.InternalResourceViewResolver">
         <property name="prefix" value="/WEB-INF/jsp/" />
@@ -66,4 +74,11 @@
         <property name="serviceInterface" value="org.onebusaway.realtime.api.VehicleLocationListener" />
     </bean>
 
+<!--    &lt;!&ndash; Security Configuration &ndash;&gt;-->
+<!--    <sec:http pattern="/swagger-resources/**" security="none"/>-->
+<!--    <sec:http pattern="/swagger-ui.html" security="none"/>-->
+<!--    <sec:http pattern="/v2/api-docs" security="none"/>-->
+<!--    <sec:http pattern="/webjars/**" security="none"/>-->
+
+<!--    <sec:global-method-security pre-post-annotations="enabled"/>-->
 </beans>
