+++ onebusaway-transit-data-federation-webapp/src/main/webapp/WEB-INF/web.xml
@@ -58,4 +58,21 @@
         <url-pattern>/remoting/*</url-pattern>
     </servlet-mapping>
 
+    <servlet-mapping>
+        <servlet-name>remoting</servlet-name>
+        <url-pattern>/swagger-ui/index.html</url-pattern>
+        <url-pattern>/swagger-ui/springfox.css</url-pattern>
+        <url-pattern>/swagger-ui/swagger-ui.css</url-pattern>
+        <url-pattern>/swagger-ui/springfox.js</url-pattern>
+        <url-pattern>/swagger-ui/swagger-ui.js</url-pattern>
+        <url-pattern>/swagger-ui/swagger-ui-bundle.js</url-pattern>
+        <url-pattern>/swagger-ui/swagger-ui-standalone-preset.js</url-pattern>
+        <url-pattern>/swagger-ui/favicon-32x32.png</url-pattern>
+        <url-pattern>/swagger-ui/favicon-16x16.png</url-pattern>
+        <url-pattern>/swagger-resources/configuration/ui</url-pattern>
+        <url-pattern>/swagger-resources/configuration/security</url-pattern>
+        <url-pattern>/swagger-resources</url-pattern>
+        <url-pattern>/v2/api-docs</url-pattern>
+    </servlet-mapping>
+
 </web-app>
